+++ NEWS.md
@@ -5,7 +5,14 @@ and aggregation to an houly axis.
 
 * Fixed bug in `pat_aggregate()` that occasionaly returned empty columns of data.
 * Added `local_examples/Jons_qc_1.R` with Jon's best take on appropriate QC of
-the hourly aggregated data.
+the hourly aggregated data. 
+* Updated AirShiny theme and include About section
+* Revamped AirShiny UI
+* Support for `pat_createAirSensor()` added to barplot 
+* Dockerized AirShiny
+* Updated Namespace
+* Added `global.R` to improve clarity of scope
+* Added hex-sticker logo for AirShiny
 
 # AirSensor 0.2.18